# Eliminate allocs in TypeConverterCache lookup path